fix: resolve feedback for variable page after review
