+++ src/app/(home-page)/github/page.jsx
@@ -7,7 +7,7 @@ import Lightning from 'components/pages/home/lightning';
 import Logos from 'components/pages/home/logos';
 import Multitenancy from 'components/pages/home/multitenancy';
 import Trusted from 'components/pages/home/trusted';
-import GetStarted from 'components/shared/get-started';
+import Cta from 'components/shared/get-started';
 import SEO_DATA from 'constants/seo-data';
 import getMetadata from 'utils/get-metadata';
 
@@ -27,7 +27,7 @@ const GithubPage = () => (
     <Multitenancy />
     <Industry />
     <Trusted />
-    <GetStarted title="Features of tomorrow.<br /> Available today." />
+    <Cta title="Features of tomorrow.<br /> Available today." />
   </>
 );
