+++ src/app/(home-page)/home/page.jsx
@@ -10,7 +10,7 @@ import Lightning from 'components/pages/home/lightning';
 import Logos from 'components/pages/home/logos';
 import Multitenancy from 'components/pages/home/multitenancy';
 import Trusted from 'components/pages/home/trusted';
-import GetStarted from 'components/shared/get-started';
+import Cta from 'components/shared/get-started';
 import SEO_DATA from 'constants/seo-data';
 import getMetadata from 'utils/get-metadata';
 
@@ -36,10 +36,7 @@ const HomePage = async () => {
       <Multitenancy />
       <Industry />
       <Trusted />
-      <GetStarted
-        title="Features of tomorrow.<br /> Available today."
-        className="pb-[307px] pt-[445px]"
-      />
+      <Cta title="Features of tomorrow.<br /> Available today." className="pb-[307px] pt-[445px]" />
     </>
   );
 };
