+++ src/components/pages/variable/bg-decor/bg-decor.jsx
