+++ src/components/pages/variable/bg-decor/index.js
