+++ src/components/pages/variable/cta/images/blue-glow-mobile.png
