+++ src/components/pages/variable/cta/images/green-glow-mobile.png
