+++ src/components/pages/variable/example/calculator/index.js
