+++ src/components/pages/variable/example/images/left-glow-mobile.png
