+++ src/components/pages/variable/example/images/right-glow-mobile.png
