+++ src/components/pages/variable/example/index.js
@@ -1,0 +1,3 @@
+import Example from './example';
+
+export default Example;
