+++ src/components/pages/variable/hero/hero.jsx
@@ -1,9 +1,9 @@
 import Container from 'components/shared/container/container';
 
 const Hero = () => (
-  <section className="hero safe-paddings relative overflow-hidden pt-[152px] xl:pt-32 lg:pt-12 md:pt-8">
+  <section className="hero safe-paddings relative overflow-hidden pt-[152px] xl:pt-32 lg:pt-12 sm:pt-9">
     <Container size="xxs">
-      <h1 className="text-[56px] font-semibold leading-dense tracking-tighter xl:text-5xl lg:text-4xl md:text-[28px]">
+      <h1 className="text-[56px] font-semibold leading-dense tracking-tighter xl:text-5xl lg:text-4xl md:text-[28px] md:leading-tight">
         Dynamically scale your
         <br /> Postgres database
       </h1>
